server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
torch>=2.1.0
diffusers>=0.32.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0
peft>=0.10.0
opencv-python-headless>=4.8.0
onnxruntime>=1.17.0
huggingface-hub>=0.23.0
numpy
truststore>=0.9.0
